ref: rename .class.ts to .ts
