+++ src/SPCPAuthClient.ts
